requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
torch>=2.0.0
torchvision>=0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0 